[RESTUCTURE] finished splitting things away from the fake backend
